Hide DB restore button when there's not a backup

+++ app/Repositories/LeafDB/LeafDBSettingRepository.php
@@ -9,41 +9,61 @@ use ErrorException;
 
 class LeafDBSettingRepository extends LeafDBConnection implements SettingRepository {
 	function backup(): bool {
-		switch (strtolower(Env::get('DB_CONNECTION'))) {
+		switch ($this->getConnectionType()) {
 			case 'sqlite':
 				$dbPath = __DIR__ . '/../SQLite/' . Env::get('DB_DATABASE');
 				parent::$db = null;;
 				copy($dbPath, "$dbPath.bak");
 				return true;
 
-				case 'mysql':
-					// TODO: Implementar respaldo de mysql
-					return false;
-
-					default: return false;
-				}
-			}
-
-			function restore(): bool {
-				switch (strtolower(Env::get('DB_CONNECTION'))) {
-					case 'sqlite':
-					$dbPath = __DIR__ . '/../SQLite/' . Env::get('DB_DATABASE');
-					parent::$db->close();
-					parent::$db = null;
-					try {
-						unlink($dbPath);
-						copy("$dbPath.bak", $dbPath);
-					} catch (ErrorException $error) {
-						Logger::log($error);
-					}
-					return true;
-
-				case 'mysql:':
-					// TODO: Implementar restauración de mysql
-					return false;
+			case 'mysql':
+				// TODO: Implementar respaldo de mysql
 				return false;
 
+			default:
+				return false;
+		}
+	}
+
+	function restore(): bool {
+		switch ($this->getConnectionType()) {
+			case 'sqlite':
+				$dbPath = __DIR__ . '/../SQLite/' . Env::get('DB_DATABASE');
+				if (file_exists("$dbPath.bak") === false) {
+					return false;
+				}
+
+				parent::$db->close();
+				parent::$db = null;
+				try {
+					unlink($dbPath);
+					copy("$dbPath.bak", $dbPath);
+				} catch (ErrorException $error) {
+					Logger::log($error);
+					return false;
+				}
+
+				return true;
+			case 'mysql:':
+				// TODO: Implementar restauración de mysql
+				return false;
+			default:
+				return false;
+		}
+	}
+
+	function hasBackup(): bool {
+		switch ($this->getConnectionType()) {
+			case 'sqlite':
+				$dbPath = __DIR__ . '/../SQLite/' . Env::get('DB_DATABASE');
+				return file_exists("$dbPath.bak");
+			case 'mysql':
+				return false;
 			default: return false;
 		}
+	}
+
+	private function getConnectionType(): string {
+		return strtolower(Env::get('DB_CONNECTION'));
 	}
 }
